Add container_tools module + function in container installation package
