+++ src/spikeinterface/sorters/__init__.py
@@ -1,4 +1,5 @@
 from .basesorter import BaseSorter
 from .sorterlist import *
-from .runsorter import *
+from .container_tools import ContainerClient, install_package_in_container
+from .runsorter import run_sorter, run_sorter_local, run_sorter_container, read_sorter_folder
 from .launcher import run_sorter_jobs, run_sorter_by_property
