+++ src/spikeinterface/sorters/tests/test_runsorter.py
@@ -1,8 +1,9 @@
 import os
 import pytest
 from pathlib import Path
+import shutil
 
-from spikeinterface import download_dataset
+from spikeinterface import download_dataset, generate_ground_truth_recording, load_extractor
 from spikeinterface.extractors import read_mearec
 from spikeinterface.sorters import run_sorter
 
@@ -15,9 +16,20 @@ else:
     cache_folder = Path("cache_folder") / "sorters"
 
 
+rec_folder = cache_folder / "recording"
+
+def setup_module():
+    if rec_folder.exists():
+        shutil.rmtree(rec_folder)
+    recording, sorting_gt = generate_ground_truth_recording(num_channels=8, durations=[10.0], seed=2205)
+    recording = recording.save(folder=rec_folder)
+
+
 def test_run_sorter_local():
-    local_path = download_dataset(remote_path="mearec/mearec_test_10s.h5")
-    recording, sorting_true = read_mearec(local_path)
+    # local_path = download_dataset(remote_path="mearec/mearec_test_10s.h5")
+    # recording, sorting_true = read_mearec(local_path)
+    recording = load_extractor(rec_folder)
+
 
     sorter_params = {"detect_threshold": 4.9}
 
@@ -37,37 +49,45 @@ def test_run_sorter_local():
 
 @pytest.mark.skipif(ON_GITHUB, reason="Docker tests don't run on github: test locally")
 def test_run_sorter_docker():
-    mearec_filename = download_dataset(remote_path="mearec/mearec_test_10s.h5", unlock=True)
+    # mearec_filename = download_dataset(remote_path="mearec/mearec_test_10s.h5", unlock=True)
+    # recording, sorting_true = read_mearec(mearec_filename)
+
     output_folder = cache_folder / "sorting_tdc_docker"
 
-    recording, sorting_true = read_mearec(mearec_filename)
+    recording = load_extractor(rec_folder)
 
     sorter_params = {"detect_threshold": 4.9}
 
-    docker_image = "spikeinterface/tridesclous-base:1.6.4-1"
+    docker_image = "spikeinterface/tridesclous-base"
 
-    sorting = run_sorter(
-        "tridesclous",
-        recording,
-        output_folder=output_folder,
-        remove_existing_folder=True,
-        delete_output_folder=False,
-        verbose=True,
-        raise_error=True,
-        docker_image=docker_image,
-        with_output=False,
-        **sorter_params,
-    )
-    assert sorting is None
+    for installation_mode in ("dev", "pypi", "github"):
+        print()
+        print("Test with installation_mode", installation_mode)
+        sorting = run_sorter(
+            "tridesclous",
+            recording,
+            output_folder=output_folder,
+            remove_existing_folder=True,
+            delete_output_folder=False,
+            verbose=True,
+            raise_error=True,
+            docker_image=docker_image,
+            with_output=False,
+            installation_mode=installation_mode,
+            **sorter_params,
+        )
+        assert sorting is None
     # TODO: Add another run with `with_output=True` and check sorting result
 
 
 @pytest.mark.skipif(ON_GITHUB, reason="Singularity tests don't run on github: test it locally")
 def test_run_sorter_singularity():
-    mearec_filename = download_dataset(remote_path="mearec/mearec_test_10s.h5", unlock=True)
-    output_folder = cache_folder / "sorting_tdc_singularity"
+    # mearec_filename = download_dataset(remote_path="mearec/mearec_test_10s.h5", unlock=True)
+    # recording, sorting_true = read_mearec(mearec_filename)
 
-    recording, sorting_true = read_mearec(mearec_filename)
+    output_folder = cache_folder / "sorting_tdc_singularity"
+    
+    recording = load_extractor(rec_folder)
 
     sorter_params = {"detect_threshold": 4.9}
 
@@ -92,6 +112,7 @@ def test_run_sorter_singularity():
 
 
 if __name__ == "__main__":
+    # setup_module()
     # test_run_sorter_local()
-    # test_run_sorter_docker()
-    test_run_sorter_singularity()
+    test_run_sorter_docker()
+    # test_run_sorter_singularity()
